requirements.txt
Flask
pymongo
mysql-connector-python
python-dotenv